# Add support for TagHelper completion item filters.