… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/PascalCasedSerializableImageElement.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/PascalCasedSerializableImageId.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSLspCompletionItem.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSLspCompletionItemIcons.cs