- Prior to this we would override TagHelper completion item "icons" by harddcoding them in our language server. These icons would end up representing roughly 40% of Razor's completion list payload and moreover wouldn't actually work through the completion filter experience. This PR changes that. The platform has expanded its set of default completion item kinds, one of which are TagHelpers, so we're now utilizing that completion item kind instead of relying on our icon setting hack that impacted performance.

![image](https://i.imgur.com/2Wr7nNP.gif)

Given this also improves perf, here's the results:

**Before**
|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |    Gen 1 |   Gen 2 |  Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|---------:|--------:|-----------:|
| 'Component Completion List Roundtrip Serialization' | 16.309 ms | 0.3178 ms | 0.3660 ms |  61.31 | 1031.2500 | 531.2500 | 31.2500 | 6514.11 KB |
|           'Component Completion List Serialization' |  2.262 ms | 0.0252 ms | 0.0236 ms | 442.06 |  128.9063 |  78.1250 | 62.5000 |   811.6 KB |
|         'Component Completion List Deserialization' | 11.911 ms | 0.2368 ms | 0.2819 ms |  83.95 |  921.8750 | 437.5000 |       - | 5702.63 KB |

**After**
|                                              Method |      Mean |     Error |    StdDev |   Op/s |    Gen 0 |    Gen 1 |   Gen 2 |  Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|---------:|---------:|--------:|-----------:|
| 'Component Completion List Roundtrip Serialization' | 11.250 ms | 0.1140 ms | 0.1066 ms |  88.89 | 718.7500 | 359.3750 |       - | 4568.71 KB |
|           'Component Completion List Serialization' |  2.086 ms | 0.0167 ms | 0.0140 ms | 479.33 |  85.9375 |  39.0625 | 19.5313 |  555.27 KB |
|         'Component Completion List Deserialization' |  9.065 ms | 0.0974 ms | 0.0911 ms | 110.32 | 640.6250 | 312.5000 |       - | 4013.45 KB |

**Results**
Roundtrip: 31% improvement
Serialization: 8% improvement
Deserialization: 24% improvement

Fixes dotnet/aspnetcore#30626

/cc @AmadeusW